+++ src/main/java/space/space_spring/service/OAuthService.java
@@ -13,6 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.LinkedMultiValueMap;
 import org.springframework.util.MultiValueMap;
 import org.springframework.web.client.RestTemplate;
+import space.space_spring.dto.jwt.TokenDTO;
 import space.space_spring.dto.jwt.TokenType;
 import space.space_spring.dto.oAuth.KakaoInfo;
 import space.space_spring.entity.User;
@@ -104,7 +105,10 @@ public class OAuthService {
         return userUtils.findOrCreateUserForOAuthInfo(email, nickname, KAKAO);
     }
 
-    public String provideJwtToOAuthUser(User userByOAuthInfo) {
-        return jwtLoginProvider.generateToken(userByOAuthInfo, TokenType.ACCESS);
+    public TokenDTO provideJwtToOAuthUser(User userByOAuthInfo) {
+        String accessToken = jwtLoginProvider.generateToken(userByOAuthInfo, TokenType.ACCESS);
+        String refreshToken = jwtLoginProvider.generateToken(userByOAuthInfo, TokenType.REFRESH);
+
+        return new TokenDTO(accessToken, refreshToken);
     }
 }
